+++ CHANGELOG.md
@@ -49,3 +49,4 @@
     - Eager Loading
     - Marker Selection Menu
 - Fix all executable-relative path issues
+- Bring back mouse support